Corrigindo int para float

+++ aula_01/src/aula_01/exc4.java
@@ -6,20 +6,20 @@ public class exc4 {
 
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
-		int n1,n2,n3,n4,diferenca;
+		
+		float n1,n2,n3,n4,diferenca;
 		System.out.println("Entre com o primeiro numero :");
-		n1 = sc.nextInt();
+		n1 = sc.nextFloat();
 		System.out.println("Entre com o segundo numero :");
-		n2 = sc.nextInt();
+		n2 = sc.nextFloat();
 		System.out.println("Entre com o terceiro numero :");
-		n3 = sc.nextInt();
+		n3 = sc.nextFloat();
 		System.out.println("Entre com o quarto numero :");
-		n4 = sc.nextInt();
+		n4 = sc.nextFloat();
 		diferenca = (n1 * n2)-(n3 * n4);
 		System.out.println("A diferença é :" + diferenca);
 		
 		
-
 	}
 
 }
